requirements.txt
tensorflow>=2.9.0
keras
numpy
pandas
matplotlib
seaborn
Pillow
streamlit>=1.20.0
scikit-learn